+++ src/editor.ts
@@ -13,6 +13,7 @@ enum Commands {
     RemoveNode = "yarnspinner.remove-node",
     ListNodes = "yarnspinner.list-nodes",
     UpdateNodeHeader = "yarnspinner.update-node-header",
+    CompileWorkspace = "yarnspinner.compile",
     ExtractString = "yarnspinner.extract",
 }