+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - A command to export a spreadsheet of all dialogue for VO recording. Can be configured in settings to control the columns, export format and a few other bits and pieces.
+- A WIP in-editor preview of yarn dialogue.
 
 ### Changed